requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-browser-reload==1.18.0
django-extensions==4.1
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.2.3
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
